# Stellify-Update/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Stellify Church Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/members/bulk-import:
    post:
      operationId: bulkImportMembers
      tags: [admin]
      summary: Bulk import members from PMDB spreadsheet data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"
  /admin/members/bulk-reset-password:
    post:
      operationId: adminBulkResetPasswords
      tags: [admin]
      summary: Reset all members' passwords to one value (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPasswordReset"
      responses:
        "200":
          description: Bulk reset result
          content:
            application/json:
              schema:
                type: object
                required: [updated]
                properties:
                  updated:
                    type: integer
  /admin/members/{id}/reset-password:
    post:
      operationId: adminResetMemberPassword
      tags: [admin]
      summary: Reset a member's password (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPasswordReset"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
                  mustChangePassword:
                    type: boolean
  /admin/members/{id}/roles:
    patch:
      operationId: updateMemberRoles
      tags: [admin]
      summary: Update a member's ministry roles array and group mappings (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RolesUpdate"
      responses:
        "200":
          description: Roles updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /admin/members/{id}/discipleship:
    patch:
      operationId: updateMemberDiscipleship
      tags: [admin]
      summary: Toggle discipleship enabled/disabled for a member (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscipleshipUpdate"
      responses:
        "200":
          description: Discipleship status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /admin/annual-reset:
    post:
      operationId: adminAnnualReset
      tags: [admin]
      summary: Re-derive stellarFragments for all members for the given year (Admin only)
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: integer
          description: Calendar year to recalculate (defaults to current year)
      responses:
        "200":
          description: Reset complete
          content:
            application/json:
              schema:
                type: object
                required: [success, year, updated]
                properties:
                  success:
                    type: boolean
                  year:
                    type: integer
                  updated:
                    type: integer
                    description: Number of members whose stats were recalculated
  /admin/stellar-status-approvals:
    get:
      operationId: listStellarApprovals
      tags: [admin]
      summary: List stellar status approval requests (Admin/Pastor)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Approval requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StellarStatusApproval"
  /admin/stellar-status-approvals/{id}/approve:
    post:
      operationId: approveStellarStatus
      tags: [admin]
      summary: Approve a stellar status change request (Pastor only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StellarStatusApproval"
  /admin/stellar-status-approvals/{id}/reject:
    post:
      operationId: rejectStellarStatus
      tags: [admin]
      summary: Reject a stellar status change request (Pastor only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StellarStatusApproval"

  # ─── GROUPS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Member:
      type: object
      required: [id, churchId, accountType, status, role, firstName, lastName, discipleshipEnabled, stellarStatus, isTestingAccount]
      properties:
        id:
          type: integer
        churchId:
          type: string
        accountType:
          type: string
          enum: [PCM, TCM]
        status:
          type: string
          enum: [Active, Deactivated, Retired, Deleted]
        role:
          type: string
          enum: [Admin, Pastor, Servant, SharingCaptain, CgAuditor, ChurchAuditor, Member]
        firstName:
          type: string
        middleName:
          type: ["string", "null"]
        lastName:
          type: string
        nickname:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
          enum: [Male, Female, null]
        birthday:
          type: ["string", "null"]
        civilStatus:
          type: ["string", "null"]
        weddingAnniversary:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        contactNumber:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]
        discipleshipEnabled:
          type: boolean
        sbgEndDate:
          type: ["string", "null"]
        stellarStatus:
          type: string
          enum: [SBG, P2S, S2B, P2G]
        pendingStellarStatus:
          type: ["string", "null"]
          description: Status awaiting pastor approval
        stellarStatusApprovedByPastor:
          type: boolean
        isTestingAccount:
          type: boolean
          description: Testing accounts are excluded from the real StellarBoard
        mustChangePassword:
          type: boolean
          description: When true, member is required to change their password on next login
        familyGroupNumber:
          type: ["string", "null"]
        familyRole:
          type: ["string", "null"]
        cgNumber:
          type: ["string", "null"]
        cgRole:
          type: ["string", "null"]
        sharingHuddleNumber:
          type: ["string", "null"]
        sharingHuddleRole:
          type: ["string", "null"]
        salvationDate:
          type: ["string", "null"]
        baptismDate:
          type: ["string", "null"]
        sharingPermissionDate:
          type: ["string", "null"]
        bearingDate:
          type: ["string", "null"]
        dedicationDate:
          type: ["string", "null"]
        ministryRoles:
          type: ["array", "null"]
          items:
            type: string
          description: Array of additional leadership roles held by this member
        roleGroupMappings:
          type: ["object", "null"]
          additionalProperties:
            type: string
          description: Maps each leadership role to its assigned group number
        foodPreferences:
          type: ["string", "null"]
        medicalCondition:
          type: ["string", "null"]
        bloodType:
          type: ["string", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactRelationship:
          type: ["string", "null"]
        emergencyContactNumber:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
          description: IANA timezone string (e.g. "Asia/Manila"). Stored by the client on login and used for server-side streak-boundary and reminder calculations.
        createdAt:
          type: string
        updatedAt:
          type: string
    MemberInput:
      type: object
      required: [churchId, accountType, firstName, lastName, password]
      properties:
        churchId:
          type: string
        accountType:
          type: string
          enum: [PCM, TCM]
        password:
          type: string
        firstName:
          type: string
        middleName:
          type: ["string", "null"]
        lastName:
          type: string
        nickname:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        birthday:
          type: ["string", "null"]
        civilStatus:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        contactNumber:
          type: ["string", "null"]
        role:
          type: string
          enum: [Admin, Pastor, Servant, SharingCaptain, CgAuditor, ChurchAuditor, Member]
        discipleshipEnabled:
          type: boolean
        sbgEndDate:
          type: ["string", "null"]
        familyGroupNumber:
          type: ["string", "null"]
        familyRole:
          type: ["string", "null"]
        cgNumber:
          type: ["string", "null"]
        cgRole:
          type: ["string", "null"]
        sharingHuddleNumber:
          type: ["string", "null"]
        sharingHuddleRole:
          type: ["string", "null"]
        salvationDate:
          type: ["string", "null"]
        baptismDate:
          type: ["string", "null"]
        stellarStatus:
          type: string
          enum: [SBG, P2S, S2B, P2G]
...
    RolesUpdate:
      type: object
      required: [ministryRoles]
      properties:
        ministryRoles:
          type: array
          items:
            type: string
          description: Full array of leadership roles for the member
        roleGroupMappings:
          type: object
          additionalProperties:
            type: string
          description: Maps each leadership role to its assigned group number
    DiscipleshipUpdate:
      type: object
      required: [discipleshipEnabled]
      properties:
        discipleshipEnabled:
          type: boolean
          description: When false, the member cannot access Disciple Hub
    AdminPasswordReset:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 6
        temporary:
          type: boolean
          description: When true, member must change this password on next login
          default: false
...
    BulkImportInput:
      type: object
      required: [members]
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/MemberInput"
    BulkImportResult:
      type: object
      required: [imported, skipped, errors]
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
...
    StellarStatusApproval:
      type: object
      required: [id, memberId, requestedStatus, requestedById, status, createdAt]
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: ["string", "null"]
        memberChurchId:
          type: ["string", "null"]
        currentStatus:
          type: ["string", "null"]
        requestedStatus:
          type: string
          enum: [SBG, P2S, S2B, P2G]
        requestedById:
          type: integer
        requestedByName:
          type: ["string", "null"]
        approvedById:
          type: ["integer", "null"]
        approvedByName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
        updatedAt:
          type: string
...